+++ library/book/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import author, genre
+# Register your models here.
+admin.site.register(author)
+admin.site.register(genre)