+++ library/library/urls.py
@@ -23,4 +23,5 @@ urlpatterns = [
     path('', include("register.urls")),
     path('user/', include("managementUser.urls")),
     path('location/', include("location.urls")),
+    path('', include("book.urls")),
 ]